Created gui window to present results. (only design)
