+++ .gitignore
@@ -1,2 +1,4 @@
 
 *.pyc
+CoughResults/Cough Sounds.wav_coughDetections.wav
+CoughResults/Cough Sounds.wav_coughDetections.txt
