+++ run_cough_detection_gui.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file 'coughv2.ui'
+# Form implementation generated from reading ui file 'cough_simplerColors.ui'
 #
 # Created by: PyQt5 UI code generator 5.9.2
 #
 # WARNING! All changes made in this file will be lost!
+
 
 from PyQt5 import QtCore, QtGui, QtWidgets
 Qt = QtCore
@@ -59,33 +60,16 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.centralwidget = QtWidgets.QWidget(MainWindow)
         self.centralwidget.setObjectName("centralwidget")
         self.frame = QtWidgets.QFrame(self.centralwidget)
-        self.frame.setGeometry(QtCore.QRect(0, 0, 531, 611))
-        self.frame.setStyleSheet("background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(9, 3, 159, 255), stop:1 rgba(3, 0, 76, 255));\n"
+        self.frame.setGeometry(QtCore.QRect(0, 0, 530, 610))
+        self.frame.setStyleSheet("background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(221, 231, 199, 255), stop:1 rgba(238, 248, 214, 255));\n"
 "\n"
 "border-radius: 10px\n"
 "")
         self.frame.setFrameShape(QtWidgets.QFrame.StyledPanel)
         self.frame.setFrameShadow(QtWidgets.QFrame.Raised)
         self.frame.setObjectName("frame")
-        self.btn_run = QtWidgets.QPushButton(self.frame)
-        self.btn_run.setGeometry(QtCore.QRect(156, 420, 220, 60))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.btn_run.setFont(font)
-        self.btn_run.setStyleSheet("QPushButton{ \n"
-"border:none; \n"
-"border-radius: 7px; \n"
-"background-color: rgb(85, 255, 127);\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: rgba(88, 255, 107, 150);\n"
-"}")
-        self.btn_run.setObjectName("btn_run")
-        self.btn_run.setEnabled(False)
-        self.btn_run.clicked.connect(self.run_main)
-
         self.frame_2 = QtWidgets.QFrame(self.frame)
-        self.frame_2.setGeometry(QtCore.QRect(10, 60, 511, 341))
+        self.frame_2.setGeometry(QtCore.QRect(20, 50, 490, 351))
         self.frame_2.setStyleSheet("background-color: rgb(227, 227, 227);border-radius:10px;")
         self.frame_2.setFrameShape(QtWidgets.QFrame.StyledPanel)
         self.frame_2.setFrameShadow(QtWidgets.QFrame.Raised)
@@ -101,13 +85,13 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.btn_dir.setStyleSheet("QPushButton{ \n"
 "border:0.5px solid grey; \n"
 "border-radius: 7px; \n"
-"background-color: rgb(85, 255, 127);\n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
 "}\n"
 "QPushButton:hover{\n"
-"background-color: rgba(85, 255, 127, 150);\n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
 "}")
         self.btn_dir.setObjectName("btn_dir")
-        self.btn_dir.clicked.connect(self.select_dir)        
+        self.btn_dir.clicked.connect(self.select_dir)                
         self.wav_folder = QtWidgets.QTextBrowser(self.frame_2)
         self.wav_folder.setGeometry(QtCore.QRect(26, 100, 271, 51))
         font = QtGui.QFont()
@@ -116,14 +100,14 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.wav_folder.setStyleSheet("background-color:rgb(255, 255, 255)")
         self.wav_folder.setObjectName("wav_folder")
         self.label_2 = QtWidgets.QLabel(self.frame_2)
-        self.label_2.setGeometry(QtCore.QRect(360, 70, 141, 31))
+        self.label_2.setGeometry(QtCore.QRect(340, 70, 141, 31))
         font = QtGui.QFont()
         font.setPointSize(9)
         font.setUnderline(True)
         self.label_2.setFont(font)
         self.label_2.setObjectName("label_2")
         self.wavs = QtWidgets.QTextBrowser(self.frame_2)
-        self.wavs.setGeometry(QtCore.QRect(26, 180, 271, 121))
+        self.wavs.setGeometry(QtCore.QRect(26, 180, 271, 141))
         font = QtGui.QFont()
         font.setPointSize(9)
         self.wavs.setFont(font)
@@ -137,6 +121,7 @@ class Ui_MainWindow(QtWidgets.QWidget):
         font.setUnderline(True)
         font.setWeight(50)
         self.label.setFont(font)
+        self.label.setStyleSheet("")
         self.label.setObjectName("label")
         self.label_3 = QtWidgets.QLabel(self.frame_2)
         self.label_3.setGeometry(QtCore.QRect(26, 160, 151, 16))
@@ -146,14 +131,14 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.label_3.setFont(font)
         self.label_3.setObjectName("label_3")
         self.num_of_threads = QtWidgets.QSpinBox(self.frame_2)
-        self.num_of_threads.setGeometry(QtCore.QRect(360, 110, 91, 41))
+        self.num_of_threads.setGeometry(QtCore.QRect(340, 103, 91, 41))
         self.num_of_threads.setStyleSheet("background-color: rgb(255, 255, 255);")
         self.num_of_threads.setMinimum(1)
         self.num_of_threads.setMaximum(20)
         self.num_of_threads.setObjectName("num_of_threads")
         self.frame_3 = QtWidgets.QFrame(self.frame)
-        self.frame_3.setGeometry(QtCore.QRect(0, 0, 530, 30))
-        self.frame_3.setStyleSheet("background-color: rgb(101, 0, 152);")
+        self.frame_3.setGeometry(QtCore.QRect(0, 0, 531, 31))
+        self.frame_3.setStyleSheet("background-color: rgba(1, 32, 15, 100)")
         self.frame_3.setFrameShape(QtWidgets.QFrame.StyledPanel)
         self.frame_3.setFrameShadow(QtWidgets.QFrame.Raised)
         self.frame_3.setObjectName("frame_3")
@@ -169,10 +154,10 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.btn_min.setStyleSheet("QPushButton{ \n"
 "border:none; \n"
 "border-radius: 7px; \n"
-"background-color: rgb(235, 157, 0);\n"
+"background-color:rgb(240, 135, 0);\n"
 "}\n"
 "QPushButton:hover{\n"
-"background-color: rgba(235, 157, 0, 150);\n"
+"background-color: rgba(240, 135, 0, 150);\n"
 "}")
         self.btn_min.setObjectName("btn_min")
         self.btn_exit = QtWidgets.QPushButton(self.frame_3)
@@ -192,29 +177,19 @@ class Ui_MainWindow(QtWidgets.QWidget):
 "background-color: rgba(235,0,0, 150);\n"
 "}")
         self.btn_exit.setObjectName("btn_exit")
-        self.btn_exit.clicked.connect(self.on_closing)
-
-        self.label_4 = QtWidgets.QLabel(self.frame_3)
-        self.label_4.setGeometry(QtCore.QRect(11, -3, 331, 41))
+        self.btn_exit.clicked.connect(self.on_closing)        
+        self.title = QtWidgets.QLabel(self.frame_3)
+        self.title.setGeometry(QtCore.QRect(11, -3, 331, 41))
         font = QtGui.QFont()
-        font.setFamily("Calibri")
+        font.setFamily("Caladea")
         font.setPointSize(14)
         font.setBold(True)
+        font.setItalic(False)
         font.setWeight(75)
-        self.label_4.setFont(font)
-        self.label_4.setObjectName("label_4")
-        self.copy_citation = QtWidgets.QPushButton(self.frame)
-        self.copy_citation.setGeometry(QtCore.QRect(400, 570, 101, 21))
-        self.copy_citation.setStyleSheet("QPushButton{ \n"
-"border:none; \n"
-"border-radius: 7px; \n"
-"background-color: rgb(200, 130, 0);\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: rgba(200, 130, 0, 150);\n"
-"}")
-        self.copy_citation.setObjectName("copy_citation")
-        self.copy_citation.clicked.connect(self.copy_citation_)
+        self.title.setFont(font)
+        self.title.setAutoFillBackground(False)
+        self.title.setStyleSheet("background-color: rgba(255, 255, 255, 0);")
+        self.title.setObjectName("title")
         self.cite_label = QtWidgets.QPlainTextEdit(self.frame)
         self.cite_label.setGeometry(QtCore.QRect(20, 500, 510, 100))
         font = QtGui.QFont()
@@ -225,11 +200,41 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.cite_label.setMouseTracking(False)
         self.cite_label.setReadOnly(True)
         self.cite_label.setObjectName("cite_label")
-        self.btn_run.raise_()
-        self.frame_2.raise_()
-        self.frame_3.raise_()
-        self.cite_label.raise_()
-        self.copy_citation.raise_()
+        self.btn_run = QtWidgets.QPushButton(self.frame)
+        self.btn_run.setGeometry(QtCore.QRect(160, 420, 225, 60))
+        font = QtGui.QFont()
+        font.setPointSize(12)
+        self.btn_run.setFont(font)
+        self.btn_run.setStyleSheet("QPushButton{ \n"
+"border:0.5px solid grey; \n"
+"border-radius: 7px; \n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
+"}\n"
+"QPushButton:hover{\n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
+"}"
+"QPushButton:disabled{\n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 50), stop:1 rgba(255, 255, 255, 50))\n;"
+"}"
+)
+        self.btn_run.setObjectName("btn_run")
+        self.btn_run.setEnabled(False)
+        self.btn_run.clicked.connect(self.run_main)        
+        self.copy_citation = QtWidgets.QPushButton(self.frame)
+        self.copy_citation.setGeometry(QtCore.QRect(390, 570, 121, 21))
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.copy_citation.setFont(font)
+        self.copy_citation.setStyleSheet("QPushButton{ \n"
+"border:0.5px solid grey; \n"
+"border-radius: 7px; \n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
+"}\n"
+"QPushButton:hover{\n"
+"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
+"}")
+        self.copy_citation.setObjectName("copy_citation")
+        self.copy_citation.clicked.connect(self.copy_citation_)
         MainWindow.setCentralWidget(self.centralwidget)
 
         self.retranslateUi(MainWindow)
@@ -238,7 +243,6 @@ class Ui_MainWindow(QtWidgets.QWidget):
     def retranslateUi(self, MainWindow):
         _translate = QtCore.QCoreApplication.translate
         MainWindow.setWindowTitle(_translate("MainWindow", "MainWindow"))
-        self.btn_run.setText(_translate("MainWindow", "Run\n(Please select directory)"))
         self.btn_dir.setText(_translate("MainWindow", "Select\n"
 "Directory"))
         self.wav_folder.setHtml(_translate("MainWindow", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.0//EN\" \"http://www.w3.org/TR/REC-html40/strict.dtd\">\n"
@@ -256,10 +260,12 @@ class Ui_MainWindow(QtWidgets.QWidget):
         self.label_3.setText(_translate("MainWindow", "Files to be processed:"))
         self.btn_min.setText(_translate("MainWindow", "_"))
         self.btn_exit.setText(_translate("MainWindow", "X"))
-        self.label_4.setText(_translate("MainWindow", "Cough Detection Tool (v 1.0)"))
-        self.copy_citation.setText(_translate("MainWindow", "Copy citation"))
+        self.title.setText(_translate("MainWindow", "Cough Detection Tool (v 1.0)"))
         self.cite_label.setPlainText(_translate("MainWindow", "Did you find this tool useful? Please cite as:\n"
 "Nikonas Simou; Nikolaos Stefanakis; Panagiotis Zervas, \"A Universal System for Cough Detection in Domestic Acoustic Environments\" in EUSIPCO 2020."))
+        self.btn_run.setText(_translate("MainWindow", "Run\n(Please Select Directory)"))
+        self.copy_citation.setText(_translate("MainWindow", "Copy citation"))
+
 
     def minimize(self):
         self.showMinimized()
@@ -272,12 +278,11 @@ class Ui_MainWindow(QtWidgets.QWidget):
             files = glob.glob(f"{folder}/**/*.wav", recursive=True)
             if len(files)>0:
                 files = "\n".join(files).replace("\\", "/").replace(folder, "./")
-                files = "Files to be processed:\n"+files
                 self.btn_run.setEnabled(True)
-                self.btn_run.setText("RUN")
+                self.btn_run.setText("Run")
             else:
-                files = "Files to be processed:\n"+"No .wav files found in selected directory."
-                self.btn_run.setText("RUN\n(Select a valid directory)")
+                files = "No .wav files found in selected directory."
+                self.btn_run.setText("Run\n(Select a valid directory)")
                 self.btn_run.setEnabled(False)
             self.wavs.setText(files)
             self.btn_dir.setText("Change\nDirectory")
